+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* Issue #1476: Fixed BigQuery MERGE statement ambiguity on overwrite when table columns are named `target` or `source`.
+
 ## 0.44.1 - 2026-03-25
 * BigQuery API has been upgraded to version 2.60.0
 * BigQuery Storage API has been upgraded to version 3.22.1